Refactor junction and coordinate handling by introducing utils and junc modules for improved code organization and functionality
